refactor: 各コンポーネントを別のディレクトリに移す
