+++ src/app/components/top/WelcomeSection.tsx
