+++ src/app/profile/page.tsx
@@ -1,7 +1,7 @@
 import Header from "../header";
-
-import CareerSection from "../components/CareerSection";
-import SkillsSection from "../components/SkillsSection";
+import ProfileDetailSection from "../components/profile/ProfileDetailSection";
+import CareerSection from "../components/profile/CareerSection";
+import SkillsSection from "../components/profile/SkillsSection";
 
 export default function AboutPage() {
   return (
@@ -10,18 +10,11 @@ export default function AboutPage() {
 
       <main className="max-w-4xl mx-auto px-4 py-12">
         <div className="bg-white rounded-lg shadow-sm p-8">
-          <section className="border-b border-gray-200 pb-12 mb-12">
-            <h1 className="text-4xl font-bold mb-8 text-gray-900">自己紹介</h1>
-            <div className="prose max-w-none prose-lg">
-              <p className="text-gray-700 leading-relaxed mb-6">
-                こんにちは、kyo1941です。
-              </p>
-              <p className="text-gray-700 leading-relaxed mb-6">
-                ここに詳細な自己紹介を記載します。
-              </p>
-            </div>
-          </section>
 
+          <section className="border-b border-gray-200 pb-12 mb-12">
+            <ProfileDetailSection />
+          </section>
+          
           <section className="border-b border-gray-200 pb-12 mb-12">
             <CareerSection />
           </section>
@@ -29,6 +22,7 @@ export default function AboutPage() {
           <section>
             <SkillsSection />
           </section>
+          
         </div>
       </main>
     </div>
